Improvements
I've been simplifying the API a lot, because I will delegate more to the
frontend than I anticipated at first. And probably it will continue in
this way of pushing more things to the P2P communication rather than
centralized in the server. However, I would like to keep some of the
functionality of a signing authority. Especially when it comes to
content tracking, we still need an orchestrator.