Introducing the realtime api
This new batch of changes does a lot of things, of course. I feel bad
for the size of these commits tbh, but in such an early stage of a
project, with so much discovery going on, it's difficult to do one thing
at a time. Most of the times I'm not even sure what I need to do.

For example for the realtime api, I've never built a P2P service service
before, so I have a lot of things to figure out on the way. Protocols,
algorithms, data structures, user interfaces... which means that I get
stuck sometimes and decide to unblock myself by doing something else
that makes me feel productive.

After a few days, or weeks, I have a mix of bug fixes, improvements, and
new features (both complete and incomplete) that makes it very difficult
to split the code. And honestly it's not worth it at this stage of the
project. As I mentioned earlier, it's still on the discovery phase.

What does that mean for a project? Well, many things. I already gave an
example of something that I did not know how to do when I embarked in
building that feature: the realtime api (AKA websockets). When I started
the project I had a much bigger signup form, with password, because I
didn't know that we could have reliable client-side key pair generation.
OpenPGP.js changed everything! Initially I still kept the password,
which was used as the key's passphrase, but then using the key meant
asking for the password multiple times, and storing it for the session
was, from a security standpoint, indistinguishible from just
autogenerating it and storing it on our side that I went back and
reworked the whole signup process again.

Enforcing a process at this point in time feels overkill, and trying to
retrofit all these changes into commits after the fact adds so much
overhead that it would, imo, hinder discovery and creativity. So here we
are, with massive commits.